refactor(world): extract World struct from lib.rs
Move the World struct, ChunkEntry, Generator enum, and all methods
from lib.rs (488 lines) to world.rs. lib.rs now contains only module
declarations and re-exports.
